requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
idna==2.9
joblib==0.15.1
redis==3.5.2
requests==2.23.0
urllib3==1.25.9
pandas==1.1.5
beautifulsoup4==4.9.3
lxml==4.6.2
yfinance==0.1.55
python-dateutil==2.8.1
PyMySQL~=1.0.2